Keep up to date with origin (#1)
* Add backup gem.

Provides an elegant DSL in Ruby for performing backups on UNIX-like
systems.

* Update README.md

FreeFileSync under Backups

* add zookeeper etcd serf consul

* Adding Centreon to Monitoring

* Update README.md: Backups > Bup

* Update README.md

* Update README.md

* Update README.md

Added Bareos to Backup section. Bareos is Bacula fork, started because open source version of Bacula almost stopped progress. Details: https://www.bareos.org/en/faq/items/why_fork.html

* Added TRK

* Update README.md

* Add NOC Project to Network Configuration Management

* Add Tsuru

* Added Squid on Web Performance category

Added Squid which is widely used as forward and reverse proxy.

* URL cleanup

* URL Cleanup

* Added Squirrelmail on Webmails and reordered Squid

* Added Redmine to ticketing systems section

* Added Seyren to monitoring tools.

* Update README.md

* Update README.md

Added Ossec.

* Tarsnap is not actually Open Source Software

The source code is available for the sole use of the tarsnap backup service,
proper. Full license text from COPYING as follows:

Unless specified otherwise in individual files or directories, the
contents of this package is covered by the following copyright, license,
and disclaimer:

Copyright 2006, 2007, 2008, 2009, 2010, 2011, 2012, 2013 Colin Percival
All rights reserved.

Redistribution and use in source and binary forms, without modification,
is permitted for the sole purpose of using the "tarsnap" backup service
provided by Colin Percival.

THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS ``AS IS'' AND
ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED.  IN NO EVENT SHALL THE AUTHOR OR CONTRIBUTORS BE LIABLE
FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS
OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
SUCH DAMAGE.

* Added couchbase to list of key/value stores

* updated url for Vlad the Deployer

* Apache Directory Studio

* update analog link

* Changed FreeFileSync url and description

* Changed ZeroMQ category and description

* Fix cobbler link

* Add Rex in Configuration Management.

* Update README.md

Add "Openshift Origin" to Cloud Computing.

* Merge branch 'ghost-patch-1'

* Minor fixes.

* Merge branch 'master' of https://github.com/jaxxstorm/awesome-sysadmin

* add pdsh

* Fixed merged conflicts

* Fixed merged conflicts